pyproject.toml
[project]
name = "ai-agent-demo"
version = "0.1.0"
description = "AI Agent demo using Flock framework"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flock-core>=0.4.524",
    "pydantic",
    "pydantic-settings",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "routers"]

[tool.uv]
dev-dependencies = []
